# Added support for extensions.