… README.md
… cmake/google/protobuf/descriptor.upb.c
… tests/conformance_upb_failures.txt
… upb/json_encode.c
… upb/msg_internal.h
… upb/msg_test.cc
… upb/msg_test.proto
… upb/reflection.c
… upb/reflection.h
… upbc/BUILD
… upbc/common.cc
… upbc/common.h
… upbc/protoc-gen-upb.cc
… upbc/protoc-gen-upbdefs.cc