This adds support for extensions across all of upb:
- binary parser
- binary serializer
- json parser
- json serializer
- generated API (in a few cases, more work is needed here)
- reflection

The core representation for extensions is an array. Arrays with linear search beat hash tables for small N, and most messages do not have a large number of extensions present.

As of this PR, upb is now passing all conformance tests!

The code size growth is very modest. The perf is neutral for parsing and a 10% improvement for serialization!

Fixes: https://github.com/protocolbuffers/upb/issues/238

```
name                                      old time/op  new time/op  delta
ArenaOneAlloc                             21.4ns ± 0%  21.0ns ± 1%   -2.15%  (p=0.000 n=12+12)
ArenaInitialBlockOneAlloc                 6.28ns ± 0%  6.29ns ± 0%   +0.09%  (p=0.000 n=12+12)
LoadDescriptor_Upb                        53.5µs ± 1%  53.5µs ± 0%     ~     (p=0.887 n=12+12)
LoadAdsDescriptor_Upb                     3.45ms ± 0%  3.54ms ± 4%   +2.63%  (p=0.000 n=12+12)
LoadDescriptor_Proto2                      239µs ± 0%   238µs ± 0%   -0.50%  (p=0.000 n=10+12)
LoadAdsDescriptor_Proto2                  14.5ms ± 0%  14.7ms ± 0%   +1.42%  (p=0.000 n=10+10)
Parse_Upb_FileDesc<UseArena,Copy>         11.0µs ± 0%  10.6µs ± 0%   -3.26%  (p=0.000 n=11+12)
Parse_Upb_FileDesc<UseArena,Alias>        9.35µs ± 0%  9.38µs ± 0%   +0.26%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<InitBlock,Copy>        10.6µs ± 0%  10.1µs ± 1%   -4.02%  (p=0.000 n=12+12)
Parse_Upb_FileDesc<InitBlock,Alias>       8.93µs ± 0%  8.91µs ± 0%   -0.18%  (p=0.000 n=12+11)
Parse_Proto2<FileDesc,NoArena,Copy>       29.4µs ± 0%  29.2µs ± 0%   -0.80%  (p=0.000 n=12+10)
Parse_Proto2<FileDesc,UseArena,Copy>      20.8µs ± 2%  20.8µs ± 2%     ~     (p=0.671 n=12+12)
Parse_Proto2<FileDesc,InitBlock,Copy>     16.6µs ± 0%  17.2µs ±12%     ~     (p=0.219 n=12+12)
Parse_Proto2<FileDescSV,InitBlock,Alias>  16.7µs ± 0%  19.3µs ±14%  +15.36%  (p=0.000 n=12+12)
SerializeDescriptor_Proto2                5.36µs ± 1%  5.34µs ± 1%     ~     (p=0.211 n=11+12)
SerializeDescriptor_Upb                   12.5µs ± 0%  11.1µs ± 0%  -10.74%  (p=0.000 n=12+12)


    FILE SIZE        VM SIZE    
 --------------  -------------- 
   +11% +1.93Ki  +9.9% +1.58Ki    upb/def.c
     +14%    +400   +14%    +400    create_msgdef
    +8.3%    +384  +8.3%    +384    _upb_symtab_addfile
     +20%    +288   +21%    +288    resolve_fielddef
    [NEW]    +150  [NEW]    +104    upb_symtab_lookupext2
    [NEW]    +123  [NEW]     +72    _upb_symtab_lookupextfield
    [NEW]    +120  [NEW]     +72    _upb_fielddef_extlayout
    [NEW]    +105  [NEW]     +56    upb_fielddef_isextension
    [NEW]    +102  [NEW]     +56    upb_fielddef_fullname
    [NEW]    +100  [NEW]     +56    upb_msgdef_fullname
    [NEW]     +66  [NEW]     +19    fill_fieldlayout.sizes
     +13%     +64   +14%     +64    _upb_symtab_loaddefinit
     +13%     +32   +15%     +32    upb_symtab_new
    +0.8%     +16  +0.8%     +16    create_fielddef
    +5.5%      +8  [ = ]       0    upb_oneofdef_name
    +6.7%      +7  [ = ]       0    upb_oneofdef_issynthetic
    +5.4%      +7  [ = ]       0    upb_oneofdef_itof
   +47% +1.25Ki   +48%   +1024    upb/msg.c
    [NEW]    +320  [NEW]    +280    _upb_extreg_add
    [NEW]    +240  [NEW]    +192    _upb_msg_getorcreateext
    [NEW]    +199  [NEW]    +160    upb_extreg_new
    [NEW]    +184  [NEW]    +144    _upb_extreg_get
    [NEW]    +176  [NEW]    +136    _upb_msg_getext
    [NEW]    +153  [NEW]    +112    _upb_msg_getexts
    +4.7%     +10  [ = ]       0    _upb_map_new
  +9.5%    +849   +12%    +368    bazel-out/k8-opt/bin/external/com_google_protobuf/src/google/protobuf/test_messages_proto2.upb.c
    [NEW]    +248  [NEW]    +224    messages_layout
    [NEW]    +159  [NEW]     +32    protobuf_test_messages_proto2_TestAllTypesProto2_MessageSetCorrectExtension1_message_set_extension_ext
    [NEW]    +159  [NEW]     +32    protobuf_test_messages_proto2_TestAllTypesProto2_MessageSetCorrectExtension2_message_set_extension_ext
    [NEW]    +111  [NEW]     +24    src_google_protobuf_test_messages_proto2_proto_upb_file_layout
    [NEW]    +106  [NEW]     +32    protobuf_test_messages_proto2_extension_int32_ext
    [NEW]     +74  [NEW]     +32    extensions_layout
    -6.2%      -8 -25.0%      -8    protobuf_test_messages_proto2_UnknownToTestAllTypes_OptionalGroup_msginit
  +2.7%    +320  +3.0%    +332    upb/decode.c
    +3.3%    +320  +3.3%    +320    decode_msg
    +3.3%     +16  +3.5%     +16    _upb_decode
     +11%      +8   +20%      +8    delim_ops
   -35.3%     -24 -50.0%     -12    upb_find_field.none
  +3.8%    +303  +6.0%    +192    bazel-out/k8-opt/bin/external/com_google_protobuf/src/google/protobuf/test_messages_proto3.upb.c
    [NEW]    +200  [NEW]    +176    messages_layout
    [NEW]    +111  [NEW]     +24    src_google_protobuf_test_messages_proto3_proto_upb_file_layout
    -7.3%      -8 -25.0%      -8    protobuf_test_messages_proto3_ForeignMessage_msginit
  +9.6%    +252  [ = ]       0    [Unmapped]
  +6.6%    +240  +8.1%    +240    upb/reflection.c
     +23%    +112   +25%    +112    upb_msg_next
     +35%     +80   +42%     +80    upb_msg_get
     +30%     +64   +36%     +64    upb_msg_set
     +20%     +48   +23%     +48    upb_msg_has
   -11.4%     -64 -12.3%     -64    upb_msg_mutable
  +3.0%    +205  +3.0%    +205    [section .rodata]
    +3.0%    +202  +3.0%    +202    [section .rodata]
     +18%      +3   +18%      +3    parse_proto.msg
   +13%    +191   +27%     +96    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/wrappers.upb.c
    [NEW]     +96  [NEW]     +72    messages_layout
    [NEW]     +95  [NEW]     +24    google_protobuf_wrappers_proto_upb_file_layout
   +19%    +174   +18%     +64    bazel-out/k8-opt/bin/external/com_google_protobuf/conformance/conformance.upb.c
    [NEW]    +102  [NEW]     +32    conformance_conformance_proto_upb_file_layout
    [NEW]     +72  [NEW]     +32    messages_layout
   +16%    +157   +24%     +64    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/struct.upb.c
    [NEW]    +101  [NEW]     +32    google_protobuf_struct_proto_upb_file_layout
    [NEW]     +56  [NEW]     +32    messages_layout
   +88%    +137  +100%     +40    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/field_mask.upb.c
    [NEW]     +97  [NEW]     +24    google_protobuf_field_mask_proto_upb_file_layout
    [NEW]     +40  [NEW]     +16    messages_layout
   +83%    +136   +83%     +40    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/timestamp.upb.c
    [NEW]     +96  [NEW]     +24    google_protobuf_timestamp_proto_upb_file_layout
    [NEW]     +40  [NEW]     +16    messages_layout
   +86%    +130   +83%     +40    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/any.upb.c
    [NEW]     +98  [NEW]     +32    google_protobuf_any_proto_upb_file_layout
    [NEW]     +32  [NEW]      +8    messages_layout
  +0.8%    +128  +0.8%    +128    upb/json_decode.c
    +7.6%    +128  +7.8%    +128    jsondec_field
   +78%    +127   +67%     +32    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/duration.upb.c
    [NEW]     +95  [NEW]     +24    google_protobuf_duration_proto_upb_file_layout
    [NEW]     +32  [NEW]      +8    messages_layout
 -13.7%     -92  [ = ]       0    [section .strtab]
    +158%     +19  [ = ]       0    completed.0
   -18.5%      -5  [ = ]       0    _GLOBAL_OFFSET_TABLE_
    [DEL]      -7  [ = ]       0    _start
   -22.8%     -99  [ = ]       0    [section .strtab]
 -12.3%     -96 -11.1%     -72    bazel-out/k8-opt/bin/external/com_google_protobuf/google/protobuf/wrappers.upbdefs.c
    [DEL]     -96  [DEL]     -72    layouts
  -0.4%    -183  -0.5%    -164    [14 Others]
  -1.7%    -208  -1.5%    -184    bazel-out/k8-opt/bin/external/com_google_protobuf/src/google/protobuf/test_messages_proto3.upbdefs.c
    -9.1%      -8 -12.5%      -8    deps
    [DEL]    -200  [DEL]    -176    layouts
  -2.3%    -248  -2.1%    -224    bazel-out/k8-opt/bin/external/com_google_protobuf/src/google/protobuf/test_messages_proto2.upbdefs.c
    [DEL]    -248  [DEL]    -224    layouts
  +3.4% +5.64Ki  +2.9% +3.75Ki    TOTAL
```