@haberman 2021-10-03
I'm not sure why the fuzz tests are suddenly failing to link against `vsnprintf`:

> bazel-out/k8-opt-exec-2B5CBBC6-ST-ab5d94bbe0bc/bin/external/com_google_protobuf/_objs/protoc_lib/php_generator.o:php_generator.cc:function google::protobuf::compiler::php::Generator::GenerateAll(std::vector<google::protobuf::FileDescriptor const*, std::allocator<google::protobuf::FileDescriptor const*> > const&, std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, google::protobuf::compiler::GeneratorContext*, std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> >*) const: error: undefined reference to 'vsnprintf.inline'

But it's not related to this PR (this particular failure isn't even code from this repo, it's from protobuf), so I'm going to go ahead and merge this.